add options

+++ scripts/simulate.sh
@@ -8,11 +8,32 @@ TSS_BIN="$HOME/workspace/tssigner/target/debug/shuttler"
 N=3
 T=2
 
-rm -rf $HOME/.shuttler*
+# User prompt if an existing local node configuration is found.
+if [ -d "$HOME/.shuttler1" ]; then
+	printf "\nAn existing folder at '%s' was found. You can choose to delete this folder and start a new shuttler with new keys. When declined, the existing local node is started. \n" "$HOME/.shuttlerN"
+	echo "Overwrite the existing configuration and start a new shuttler cluster? [y/n]"
+	read -r overwrite
+else
+	overwrite="Y"
+fi
 
-for i in $(seq 1 $N); do
-  $TSS_BIN --home $HOME/.shuttler$i init --network testnet --port "535$i"
-done
+if [[ $overwrite == "y" || $overwrite == "Y" ]]; then
+
+  rm -rf $HOME/.shuttler*
+
+  # Init shuttler home
+  for i in $(seq 1 $N); do
+    $TSS_BIN --home $HOME/.shuttler$i init --network testnet --port "535$i"
+    sed -i -e "/rpc_address =/ s/= .*/= \"127.0.0.1:818$i\"/" $HOME/.shuttler$i/config.toml
+  done
+
+  # fund relayer address
+  for i in $(seq 1 $N); do
+    echo "\n *** Fund relayer $i ***"
+    $BINARY tx bank send validator $($TSS_BIN --home "$HOME/.shuttler$i" address | grep tb1) 10000000uside --chain-id $CHAINID --keyring-backend $KEYRING --yes --fees 2000uside
+    sleep 6
+  done
+fi
 
 declare -a TSS_PARTICIPANTS
 for i in $(seq 1 $N); do
@@ -63,7 +84,7 @@ echo $DCA_PROPOSAL > ../build/dca.json
 cat ../build/dca.json
 $BINARY tx gov submit-proposal ../build/dca.json --from validator --fees 1000uside --chain-id $CHAINID --keyring-backend $KEYRING -y
 
-sleep 10
+sleep 6
 
 # Vote active proposals
 $BINARY q gov proposals --output json | jq -r '.proposals| .[] | select(.status == 2) | .id'| while read -r id; do 
@@ -73,7 +94,8 @@ $BINARY q gov proposals --output json | jq -r '.proposals| .[] | select(.status 
 done
 
 for i in $(seq 1 $N); do
-  $TSS_BIN --home $HOME/.shuttler$i start > $HOME/.shuttler$i/output.log &
+  # RUST_BACKTRACE=1 
+  $TSS_BIN --home $HOME/.shuttler$i start --oracle --agency > $HOME/.shuttler$i/output.log &
 done
 
-tail -f $HOME/.shuttler$i/output.log
+tail -f $HOME/.shuttler1/output.log
